refactor: enhance file names and packages
