+++ app/src/main/java/com/london/tudee/presentation/screens/home/HomeInteractions.kt
@@ -1,0 +1,28 @@
+package com.london.tudee.presentation.screens.home
+
+import com.london.tudee.domain.entities.Category
+import com.london.tudee.domain.entities.Priority
+import com.london.tudee.domain.entities.Task
+import com.london.tudee.presentation.screens.task.TaskModifyInteractions
+
+interface HomeInteractions : TaskModifyInteractions {
+    override fun showDatePicker()
+    override fun hideDatePicker()
+    fun showBottomSheet()
+    override fun hideBottomSheet()
+    fun onThemeSwitched(isDarkMode: Boolean)
+    fun loadCategories()
+    fun validateForm()
+    fun clearMessages()
+    fun loadTask(task: Task)
+    fun onClickMove()
+    fun showTaskDetailsBottomSheet()
+    fun hideTaskDetailsBottomSheet()
+    fun onEditTask(taskId: Int)
+    override fun updateTitle(title: String)
+    override fun updateDescription(description: String)
+    override fun updateDate(date: Long)
+    override fun updatePriority(priority: Priority)
+    override fun updateCategory(category: Category)
+    override fun saveTask()
+}
