+++ app/src/main/java/com/london/tudee/presentation/screens/home/HomeViewModel.kt
@@ -10,9 +10,8 @@ import com.london.tudee.domain.entities.TaskStatus
 import com.london.tudee.domain.services.AppPreferencesService
 import com.london.tudee.domain.services.CategoryService
 import com.london.tudee.domain.services.TaskService
-import com.london.tudee.presentation.base.HomeInteractions
-import com.london.tudee.presentation.screens.task.add_edit_task_bottom_sheet.AddOrEditTaskUiState
-import com.london.tudee.presentation.screens.task.taskdetails.TaskDetailsBottomSheetUiState
+import com.london.tudee.presentation.screens.task.task_details.TaskDetailsBottomSheetUiState
+import com.london.tudee.presentation.screens.task.task_modify.TaskModifyUiState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -33,7 +32,7 @@ class HomeViewModel(
     private val _uiState = MutableStateFlow(HomeUiState())
     val uiState = _uiState.asStateFlow()
 
-    private val _taskUiState = MutableStateFlow(AddOrEditTaskUiState())
+    private val _taskUiState = MutableStateFlow(TaskModifyUiState())
     val taskUiState = _taskUiState.asStateFlow()
 
     init {
@@ -213,7 +212,7 @@ class HomeViewModel(
         }
     }
 
-    override fun initializeForEdit(taskId: Int) {
+    override fun onEditTask(taskId: Int) {
         viewModelScope.launch(Dispatchers.IO) {
             _taskUiState.update { it.copy(isLoading = true) }
 
@@ -259,17 +258,17 @@ class HomeViewModel(
         _taskUiState.update { it.copy(description = description) }
     }
 
-    override fun updateSelectedDate(date: Long) {
+    override fun updateDate(date: Long) {
         _taskUiState.update { it.copy(selectedDate = date) }
         validateForm()
     }
 
-    override fun updateSelectedPriority(priority: Priority) {
+    override fun updatePriority(priority: Priority) {
         _taskUiState.update { it.copy(selectedPriority = priority) }
         validateForm()
     }
 
-    override fun updateSelectedCategory(category: Category) {
+    override fun updateCategory(category: Category) {
         _taskUiState.update { it.copy(selectedCategory = category) }
         validateForm()
     }
