+++ app/src/main/java/com/london/tudee/presentation/screens/task/task_modify/FakeData.kt
@@ -1,0 +1,88 @@
+package com.london.tudee.presentation.screens.task.task_modify
+
+import com.london.tudee.domain.entities.Category
+
+val category = listOf(
+    Category(
+        id = 1,
+        title = "Education",
+        // arName = "التعليم",
+        iconRes = "",
+        isDefault = true,
+        taskCount = 0,
+        // tint = primaryColor
+    ),
+    Category(
+        id = 2,
+        title = "Shopping",
+        // arName = "التسوق",
+        iconRes = "",
+        isDefault = true,
+        taskCount = 0,
+        // tint = secondaryColor
+    ),
+    Category(
+        id = 3,
+        title = "Medical",
+        // arName = "طبي",
+        iconRes = "",
+        isDefault = true,
+        taskCount = 0,
+        // tint =primaryColor
+    ),
+    Category(
+        id = 4,
+        title = "Gym",
+        // arName = "رياضة",
+        iconRes = "",
+        isDefault = false,
+        taskCount = 0,
+        //   tint = primaryColor
+    ),
+    Category(
+        id = 5,
+        title = "Entertainment",
+        // arName = "ترفيه",
+        iconRes = "",
+        isDefault = false,
+        taskCount = 0,
+        //tint = primaryColor
+
+    ),
+    Category(
+        id = 6,
+        title = "Cooking",
+        // arName = "طبخ",
+        iconRes = "",
+        isDefault = false,
+        taskCount = 0,
+        // tint = primaryColor
+    ),
+    Category(
+        id = 7,
+        title = "Family & Friends",
+        // arName = "العائلة والأصدقاء",
+        iconRes = "",
+        isDefault = false,
+        taskCount = 0,
+        // tint = primaryColor
+    ),
+    Category(
+        id = 8,
+        title = "Traveling",
+        // arName = "سفر",
+        iconRes = "",
+        isDefault = false,
+        taskCount = 0,
+        // tint = primaryColor
+    ),
+    Category(
+        id = 9,
+        title = "Agriculture",
+        //arName = "زراعة",
+        iconRes = "",
+        isDefault = false,
+        taskCount = 0,
+        //   tint = primaryColor
+    )
+)
