+++ app/src/main/java/com/london/tudee/presentation/utils/TasksScreenUtils.kt
@@ -1,4 +1,4 @@
-package com.london.tudee.presentation.screens.tasks
+package com.london.tudee.presentation.utils
 
 import com.london.tudee.presentation.utils.DateFormatter.toYear
 import kotlinx.datetime.DatePeriod
@@ -10,8 +10,8 @@ import kotlinx.datetime.plus
 
 object TasksScreenUtils {
 
-     fun LocalDate.lengthOfMonth(date: Long) =
-         when (month) {
+    fun LocalDate.lengthOfMonth(date: Long) =
+        when (month) {
             Month.FEBRUARY -> {
                 if (isLeapYear(date.toYear().toInt())) 29 else 28
             }
@@ -19,14 +19,15 @@ object TasksScreenUtils {
             Month.APRIL, Month.JUNE, Month.SEPTEMBER, Month.NOVEMBER -> {
                 30
             }
+
             else -> 31
         }
 
-     fun isLeapYear(year: Int): Boolean =
+    fun isLeapYear(year: Int): Boolean =
         (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0)
 
-     fun getDayRangeMillis(targetDate: LocalDate): Pair<Long, Long> {
-        val timeZone = TimeZone.currentSystemDefault()
+    fun getDayRangeMillis(targetDate: LocalDate): Pair<Long, Long> {
+        val timeZone = TimeZone.Companion.currentSystemDefault()
         val startOfDayMillis = targetDate.atStartOfDayIn(timeZone).toEpochMilliseconds()
         val endOfDayMillis =
             targetDate.plus(DatePeriod(days = 1)).atStartOfDayIn(timeZone).toEpochMilliseconds() - 1
