+++ app/src/main/res/values-ar/strings.xml
@@ -99,4 +99,5 @@
     <string name="loading">جاري التحميل....</string>
     <string name="there_was_an_unexpected_error">هناك خطا غير متوقع</string>
     <string name="tasks">المهام</string>
+    <string name="empty_task">Empty Task</string>
 </resources>
