+++ app/src/main/res/values/strings.xml
@@ -106,4 +106,5 @@
     <string name="tasks">Tasks</string>
     <string name="delete_category_successfully">Delete Category Successfully</string>
     <string name="edit_category_successfully">Edit Category Successfully</string>
+    <string name="empty_task">Empty Task</string>
 </resources>
